Issue 3065 - New discourse module based on docker
Introduce a new module for installing discourse based on their
recommended Beginner Docker install guide [1].

The next steps are to run a community forum based on this module, and
later on migrate the intraforum to run using the same module.

The patch assumes that it is run on a base-image that supports modern
versions of docker (i.e. Debian Jessie).

Note that this patch is a subset of the work Wladimir did to
upgrade to the current way of installing the intraforum [2].

[1] https://github.com/discourse/discourse/blob/master/docs/INSTALL.md
[2] https://codereview.adblockplus.org/29347356/